MNT: Move most of __main__.py into utils.py
